Fallback to activation if old manifest doesn't exist.
I've gotten tired of writing bash conditionals in every place
I use `smfh` since I need to account for the possibility
that it's the first time and that there's no old manifest to
diff with. This change essentially just moves that conditional
into smfh itself.

The lack of proper error handling in `Manifest::read` means
that this is vulnerable to time-of-check to time-of-use bugs,
though these should be unlikely when `smfh` itself is the
one making the changes.